Proyecto 11 Pokemon/etl-pokeapi/requirements.txt
requests==2.32.5
python-dotenv==1.2.2
matplotlib==3.10.8
openpyxl==3.1.5
psycopg2-binary==2.9.11
SQLAlchemy==2.0.48
alembic==1.18.4
numpy==2.4.2
plotly==6.5.2
streamlit==1.54.0
streamlit-lottie==0.0.5
pandas==2.3.3